Day 8/requirements.txt
streamlit==1.32.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-tavily>=0.0.1
langgraph>=0.0.20
langgraph-checkpoint-sqlite>=0.0.1
tavily-python>=0.3.0
python-dotenv>=1.0.0
PyPDF2==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
google-generativeai>=0.3.0 